+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## 2017-07-23
 - First release: Merge folders from local disks and from windows/samba shares
+- Load and correct single values at start